Buildfix for broken hw842_(de)compress from my previous commit which changed most types to size_t.

+++ include/hw842.h
@@ -3,10 +3,10 @@
 
 #include <stdint.h>
 
-int hw842_compress(const uint8_t *in, unsigned int ilen,
-		   uint8_t *out, unsigned int *olen);
+int hw842_compress(const uint8_t *in, size_t ilen,
+		   uint8_t *out, size_t *olen);
 
-int hw842_decompress(const uint8_t *in, unsigned int ilen,
-		     uint8_t *out, unsigned int *olen);
+int hw842_decompress(const uint8_t *in, size_t ilen,
+		     uint8_t *out, size_t *olen);
 
 #endif
